requirements.txt
aiofiles==24.1.0
astroid==3.3.7
bcrypt==4.2.1
better-profanity==0.7.0
bidict==0.23.1
certifi==2024.8.30
dill==0.3.9
dnspython==2.7.0
eventlet==0.38.0
greenlet==3.1.1
h11==0.14.0
html5tagger==1.3.0
httptools==0.6.4
iniconfig==2.0.0
isort==5.13.2
mccabe==0.7.0
multidict==6.1.0
packaging==24.2
platformdirs==4.3.6
pluggy==1.5.0
pymongo==4.10.1
pytest==8.3.4
python-engineio==4.10.1
python-socketio==5.11.4
PyYAML==6.0.2
sanic==24.6.0
Sanic-Cors==2.2.0
sanic-ext==23.12.0
sanic-routing==23.12.0
setuptools==75.6.0
simple-websocket==1.1.0
tomlkit==0.13.2
tracerite==1.1.1
typing_extensions==4.12.2
ujson==5.10.0
uvloop==0.21.0
websockets==14.1
wsproto==1.2.0
